requirements.txt
certifi==2018.1.18
Django==2.0.4
djangorestframework==3.8.2
Markdown==2.6.11
mysqlclient==1.3.12
pytz==2018.4
